Win32: Add macros for common functions like strdup
For some reason, Microsoft have an aversion to including functions like
strdup, open, read, write, and close that are part of the POSIX
standard, but still keep them present with a _ prefix.

This commit adds #defines for some of these functions so that we can
keep using the existing code without having to add all lots of ugly
underscores everywhere.