Win32: Don't misuse TEMP environment variable
In dfutil's convertFile function, there is a facility to allow the user
to explicitly specify a temporary directory to be used, instead of a
randomly-generated name. This done through the TEMP environment
variable, which is not normally set under OS X or Linux, but is on
windows.

To avoid confusion, look for the DFUTIL_TEMP environment instead.