Win32: Fix temp dir creation in dfutil
Win32 lacks the mkdtemp function, which we normally use in dfutil to
create a temporary directory for conversion. When running on windows,
instead create a directory with the fixed name "dfutil.temp" (unless the
DFUTIL_TEMP environment variable is set to something else).

We don't really need a random name; as long as the directory is cleared
before conversion, it will work fine.

Ideally it would be good to have a way to convert files without
requiring a temporary directory, but that's something to be addressed
separately.