Reorganise directory hierarchy
Change the directory structure of the project to cleanly separate
different modules. The module hierarchy is as follows:

Socrates
 |- consumers
 |   |- dfconvert
 |   \- dfutil
 \- DocFormats
     |- 3rdparty
     |- api
     |- core
     \- filters
         |- HTML
         |- LaTeX
         |- ODF
         \- OOXML

Each module has a src and tests directory, and (in some cases) a headers
directory. In the case of the filters module, we have one headers
directory which is shared between all of the individual filters (though
this is currently empty).