Xcode: Avoid warning about missing library symbols
When Xcode compiles a library and notices that it does not contain any
symbols, it gives the following warning:

    warning for library: .../libtest_html.a the table of contents is
    empty (no object file members in the library define global symbols)

To avoid this, we define an empty "stub" function in each of the single
source files for these new libraries. These functions only need to
remain as long as there is nothing else in the library, to avoid the
spurious warnings - once these libraries start getting implementation
code in them, the stub functions can be removed.