dfutil: Fix build on mac
dfutil currently includes two Objective C files, FunctionTests.m and
StringTests.m. These ultimately need to be replaced with C alternatives,
but currently they work by testing certain things and making sure that
the results exactly match the results produced by Apple's equivalent of
the relevant functions.

Because these files can only be compiled on iOS/OS X, they need to be
conditionally included in the list of source files for dfutil. This
involves adding back the platform test, and defining the variable
SOURCES to be the list of source files, and then using that in the
add_executable command. The equivalent line from the top-level
CMakeLists.txt has now been removed, as this is only relevant to dfutil,
not the project as a whole.

This change is a temporary measure necessary to build on Apple
platforms. Ultimately, we'll be able to switch back to defining the list
of source files for dfutil directly, without any platform-dependent
build configuration.

Apart from this, the LIBS variable has also been added to the target
link libraries. This is because on iOS and OS X, DocFormats depends on
the Foundation and ImageIO frameworks for determining image dimensions.
Other platforms will use different libraries, but these will be set in
the LIBS directory at the top level as is currently done for Apple
platforms.