Move all Word headers out of core
There was a single function call, WordSetupTableGridStyle, made from
CSSClassNames.c, which resided in core/src/css. The latter file included
some headers from the Word filter, which included more headers, and so
forth, which all had to be brought into core to get it to build
independently.

Fortunately however, the functions in CSSClassNames.c are only actually
used in the Word filter, so we can simply move it into there. The
algorithms implemented in CSSClassNames.c are generally applicable to
other filters, but currently contain a call to WordSetupTableGridStyle;
once this is abstracted away, and the need arises to use these functions
frmo other filters, we can move the file back into core.