Use square images in all tests
On non-Apple platforms, we don't currently have implementations of the
DFGetImageDimensions function, which is used to determine the aspect
ratio of an image file when only one of the dimensions in the HTML file
being converted is specified (e.g. UX Write typically just sets the
width, leaving the height to be be automatically computed from the image
aspect ratio).

About 35 tests have been failing on other platforms because of this.
While we ultimately require tests for filling in the missing dimension
of an image when updating a word document, as an interim solution we'll
just work on the assumption that if we're unable to get the dimensions
of an image, we can treat it as 1:1.

The necessary changes to make the tests actually work on other platforms
have not yet been made, but this at least ensures that all tests are
using images which do actually have a 1:1 dimension (specifically, a
64x64 png), so that when run with on Apple platforms the results are the
same as what we'll get on those that don't yet have the image dimension
function implemented.