All tests now pass on OS X, Linux, and Windows
Ignore errors reported by DFGetImageDimensions, assuming an image size
of 0x0 when this occurs. When something needs to be done based on the
aspect ratio of the image, it is taken to be 1:1.

This change is to cater for the fact that DFGetImageDimensions is not
currently implemented on Linux and Windows. While this ultimately needs
to be fixed so we can properly handle non-square images in cases where a
HTML specifies a width but not a height, it at least means that we now
have all of the tests passing successfully on the three platforms.