DFAddDirContents: Don't use DFArray or DFError
Avoid using DFArray in DFAddDirContents, as a step towards removing the
dependency of the platform module upon core. We instead use the
DFDirEntryList type, which is simply a linked list of string values
which is manipulated manually, without requiring any convenience
functions defined in core.

This function also no longer takes a DFError object as a paramater,
though others in DFPlatform.h still do.