Don't store bin/lib/include dirs in repository
These are only intended to contain relevant files from the downloaded
binary distributions, and excluding them from the repository makes it
simpler to blow them away and replace them with those created by the
script when necessary.

The unistd.h file that was present in the include does not necessary,
because it was only the for zconf.h, but zconf.h does not attempt to
include unistd.h if HAVE_UNISTD_H is not set, as is the case on Windows.