Add DFextZipAppendNewFile
A handle to a zip file is either in read-only mode, or write-only mode
(as determined by the zipFlag field of the object). In the former case,
DFextZipOpenNextFile opens an existing file for reading at the next
position in the file. In the latter case, it created a new file at the
end of the zip file for writing.

When reading the next entry, the entryName parameter must be writable -
that's how the name is passed back to the caller. As such, the caller
provides a writable string, consisting of a char * and a maximum length.
In this case, the parameter types are ok.

When writing the next entry, the entryName parameter plays the opposite
role - it is the caller telling the function what name to give to the
new file. In what is currently the only place where this is called,
zipAddFile in DFZipFile.c, the filename is a read-only string (const
char *). Although the DFextZipOpenNextFile didn't actually overwrite the
string in the case of a write-only zip stream, the compiler would warn
that a value of const char * was being supplied to a function expecting
a char * value (that is, a read-only string being supplied where a
writable string parameter was expected).

This commit splits DFextZipOpenNextFile into two functions -
DFextZipOpenNextFile for reading and DFextZipAppendNewFile for writing.
The latter accepts a read-only string, and thus doesn't generate a
compiler warning.