DFUnzip: Fix warning about unsigned char * cast
The DFextZipReadCurrentFile function expected a char * parameter, but we
were passing in an unsigned char *, resulting in a compiler warning.

This commit changes DFextZipReadCurrentFile to instead take a void *
parameter, and similarly DFextZipWriteCurrentFile to take a const void *
parameter. This allows any pointer types to be used with the functions.