dfutil: Fix pretty-printing option (-pp)
Word_toPlainFromDir was incorrectly interpreting relative path names in
OPC relationship files for word documents. Instead of calculating the
path relative to the root of the package, it would use the path relative
to the location of the document.xml file.

This was causing the pretty-printing routine to encounter an error when
looking for files like styles.xml, because it would check for them in
the root of the package, rather than the directory in which they were
located (usually /word).