Get rid of DFCreateTempDir
For a number of operations, like running tests and converting documents,
we used to require the existence of a temporary directory into which we
could extract the contents of a .docx or other zip-based package.
However, since the implementation of the DFStorage interface, with both
an in-memory implementation, and a zip-based implementation, it's no
longer necessary to write the XML files from a zip package out to the
filesystem.

DFCreateTempDir used to rely on mkdtemp, which generates temporary
directories in a secure manner (that is, avoiding a race condition
between the directory's creation time and the application getting
exclusive access to it). However, this function is not available on
windows, leaving us with the choice of having an #ifdef WIN32 to use a
fixed directory name on windows, or doing the same thing on all
platforms. Compiling on Linux would give a warning about the mktemp
function being unsafe.

Given that we don't actually need temporary directories any more for
testing or conversion though, we can just remove this function
altogethter, avoiding this dilemma.

The one place where temporary files are still used is when dftest is run
with the -diff option, which prints out the differences between the
expected and actual results of a test case if there is a failure. This
has been changed to use the fixed filenames 'dftest-diff-from.tmp' and
'dftest-diff-to.tmp' so it can run the diff command as a sub-process.

Another nice advantage of this is when dftest or dfutil crashes, you
don't get a dfutil.XXXXXX directory lying around.