Word: Move WordConverter{Get,Put}2 code
Previously, the functions WordConverterGet and WordConverterPut would
simply create a WordConverter object and then call through to the '2'
version of the same function.

Given that there's little in the way of error handling that needs to
happen in the latter, it makes sense to do it all in one function.