Word: Remove idPrefix parameter
It used to be that code using the DocFormats library would supply a
prefix to be used on all id attributes in the HTML file which were used
to maintain a mapping back to the elements in the original document.
The reason for this was that in UX Write, if you copied & pasted from
one word document to another, the HTML code would be pasted as-is, and
this could result in incorrect mappings being stored in the destination
file.

UX Write handles this differently now (can't remember the details; it
was a long time ago I made the change), and for a long time we've been
using the fixed prefix of "word" for all id attributes. So we don't need
that as a parameter to the conversion functions.