Cursor hit testing inside empty footnotes/endnotes
Position_atPoint mostly relies on document.caretRangeFromPoint for hit
testing, but this doesn't work for empty footnotes or endnotes, as there
is no actual text. Instead, caretRangeFromPoint returns the position
directly after the last letter before the note. So we do the same thing
as for images, which is checking if the adjacent image or empty note has
a bounding rect containing the specified point.