Cursor hit testing after footnotes/endnotes
Previously, Position_atPoint would return the position directly before
an empty footnote or endnote if called with a point that was inside the
first character following it. We now check to see if the x position is
to the right of the empty note, and if so, returns the position directly
before the first character after the note.