Fix cursor position in empty caption
This commit fixes a bug where if you place the cursor inside an empty
figure or table caption, it's displayed at the left of the screen,
instead of directly after the "Table X: " or "Figure X: " generated
content.

We can't directly get the bounding rect of generated content, so we
temporarily insert a text node containing a single space character, get
the position to the right of that character, and then remove the text
node. This allows us to show the user the position at which text will be
inserted if they start entering text in the caption.