Tables: Make row/column addition work at edges
Modify addAdjacentRow and addAdjacentColumn such that they work if the
cursor is placed immediately before or immediately after a table. This
is necessary for the user to be able to add a column to the left of the
table, or a row to the top of the table, since in all other cases they
are added to the right or below the selection.