Sample code showing how to implement bindings
This code comes from UX Write and is what provides the glue between
Objective C and Javascript. It can't actually be used in any meaningful
way without the rest of the UX Write source code (which is not part of
the repository), but I've put it here to aid in the implementation of
C++ bindings in the Qt app, as well as any other languages for which
bindings are to be implemented.