Desktop app: Basic infrastructure for Editor lib
This commit sets up the basic architecture for a functional editor
implemented as a Qt app. While it currently does not have the necessary
facilities to actually edit documents, it lays the groundwork by
providing a web view injected with the javascript editing library, and a
C++ interface to interact with this library.

See JSInterface.h for information about how the interaction between the
two languages works. Essentially, this file contains a series of classes
containing methods which call through to the corresponding functions in
javascript. These methods handle marshalling of arguments from Qt types
to JSON types, and similarly in the other direction for return values.

A callback interface is also provided to allow JS code to cause C++
functions to be invoked in response to certain events occurring, like
the cursor moving (and thus requiring redisplay).

The user interface is very bare-bones and incomplete; it simply contains
a web view, loaded at startup with a sample document, and buttons to
invoke various functions in the editor library like inserting/deleting
characters and moving the cursor. These are intended not for actual
usage, but as demonstrations of how to invoke JS methods from Qt even
callbacks ("slots"). Functionality to open and save documents is also
absent.

Currently this only builds on OS X, but that's an issue with the CMake
config files which, once fixed, should enable building on Linux and
Windows as well.