Use document hash to protect against misuse of put
The put operation was potentially quite dangerous if you didn't know
what you were doing, or just accidentally used it with the wrong source
document and/or HTML file. The algorithms used to update an existing
Word document (and, in the future, other formats) depend on a correct
mapping between HTML elements and the XML elements in the source
document from which they were generated. If the wrong HTML file was used
to update a .docx file, an incorrect mapping could lead to file
corruption in the sense of the XML data not matching the expecetd
schema.

To protect against this, in the get operation, we now compute a hash of
all the XML documents inside a .docx package, and store this in a <meta>
tag in the HTML document. In the put operation, we re-compute this hash,
and check that it matches the one given in the <meta> tag of the HTML
document being used to update the source.

A common situation where this protection helps is if you try to invoke
put twice on the same pair of files. For example:

    # dfconvert get test.docx test.html
    ... Make some changes to test.html ...
    # dfconvert put test.docx test.html
    # dfconvert put test.docx test.html

The second put will now fail, because the first changed the XML
structure of the docx file in a way that made the node identifiers
referenced from the HTML id attributes no longer match in the intended
way. Now, we detect that on th second put, the hash of test.docx will be
different to that given in test.html, and we abort with an error message
stating: "HTML document was generated from a different file to the one
being updated".

To avoid having to update all the test cases, we strip out the hash when
doing a test of get, and set the special value "ignore" in the <meta>
tag before doing a put, which disables the check.