Merge EditorPrivate class into Editor
The Editor class, as exposed publicly, hid most of its details in a
separate EditorPrivate class defined inside Editor.cpp. There's going to
be a *lot* of work done on the Editor class in the near future, and I
expect there'll be need to access some of the properties/methods from
other files, so I've merged the two classes together for now.

It's a common idiom in Qt to have a separate private class for details
that you don't want exposed generally, and we'll likely want this for
Editor in the future, but it's more convenient to have it all in the one
class for now.