Added editorFramework in experiments.
The editorFramework is the implementation of our discussions and several experiments.

It contains a layered model, in order to isolate the external toolkit (different license) from the functionality of the system
It contains provisions for 2 toolkits: Qt and python bindings for a web server version.

This is the initial framework, code will be copied from other experiments and added slowly.