Add Objective C framework code from UX Write
These files come from the portion of the UX Write editor that works on
both Apple platforms (that is, it can run on either OS X or iOS). It's
in the repository for illustrative purposes only, to assist with the
creation of the framework for the Corinthia editor UI. The code does not
compile independently in its present form.