Change uxwrite- CSS class prefix to corinthia-
There are a number of cases both in the editor and in DocFormats where
we use "special" classes that have a specific meaning within the context
of Corinthia. For historical reasons, these were prefixed with the
string "uxwrite-". This commit changes all source files and tests to use
the prefix "corinthia-" instead.