Fix crash on get with invalid list numbering id
Gracefully handle the case where a paragraph has an ilvl or numId
property that does not correspond to an entry in numbering.xml.
Technically this is only the case for invalid documents, but some users
have reported this problem occurring for documents created with the iOS
version of Pages, which may not be handling numbering correctly.