Fix crash on null namespace prefix
When parsing an XML file and encountering a namespace we do not already
know about and that has a null prefix, we would previously crash, due to
the call to xstrdup(prefix) on a NULL pointer in DFNamespaceInfoNew.

It's valid for a namespace to have no prefix, so change the xstrdup
calls in this function to DFStrDup, which handles NULL strings.